requirements.txt
yt-dlp>=2024.3.10
openai-whisper==20240930
ffmpeg-python>=0.2.0
requests
